fix touch event handling: touch points are not scaled twice (don't override in MonocleView) gestures are now called with logical pixels instead of physicals. GlassViewEventHandler will do the conversion.

+++ modules/graphics/src/main/java/com/sun/javafx/tk/quantum/GestureRecognizers.java
@@ -66,7 +66,7 @@ class GestureRecognizers implements GestureRecognizer {
     }
 
     public void notifyNextTouchEvent(long time, int type, long touchId,
-                                     int x, int y, int xAbs, int yAbs)
+                                     double x, double y, double xAbs, double yAbs)
     {
         final GestureRecognizer[] wl = synchWorkList();
         for (int idx = 0; idx != wl.length; ++idx) {
